+++ bakedanuki/bakedanuki-tools/python/bd_tools/channel_editor/__init__.py
@@ -1,6 +1,7 @@
 # coding: utf-8
 """選択ノードのbool・float系属性を入力するツール。"""
 
+from . import config
 from .ui import (
     ChannelEditorWindow,
     WORKSPACE_CONTROL_NAME,
@@ -12,6 +13,7 @@ from .ui import (
 )
 
 __all__ = [
+    "config",
     "ChannelEditorWindow",
     "WORKSPACE_CONTROL_NAME",
     "close",